Reference the todo index in toggleTodo correctly
Previously we were attempting to get the index from the e.target.dataset object but that object is empty when we receive it in toggleTodo. Instead, if you look at the template we see that both the event object and the index is passed onto toggleTodo. Therefore, we only need the index to reference the correct index of the todo to toggle.